requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
crispy-bootstrap5==2023.10
Django==5.0b1
django-crispy-forms==2.1
django-rename-app==0.1.7
djangorestframework==3.14.0
psycopg==3.1.12
psycopg-binary==3.1.12
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
